fix: 회원가입 API를 같이 만들어 일어나는 오류 해결 완료
