+++ server/src/main/java/com/team_60/Mocco/member/entity/Member.java
@@ -43,7 +43,7 @@ public class Member extends Auditable {
     private String providerId;
 
     @Column
-    private String roles;
+    private String roles = "ROLE_USER";
 
     public List<String> getRoleList() {
         if(this.roles.length() > 0){
